simpleagent/requirements.txt
# requirements.txt
python-dotenv
pyyaml
requests
pyinstaller
